requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile --all-extras
#
aiohttp==3.8.4
    # via cloud-sql-python-connector
aiosignal==1.3.1
    # via aiohttp
asn1crypto==1.5.1
    # via scramp
astropy==5.2.1
    # via mpcq (setup.py)
async-timeout==4.0.2
    # via aiohttp
attrs==22.2.0
    # via
    #   aiohttp
    #   pytest
cachetools==5.3.0
    # via google-auth
certifi==2022.12.7
    # via requests
cffi==1.15.1
    # via cryptography
cfgv==3.3.1
    # via pre-commit
charset-normalizer==3.1.0
    # via
    #   aiohttp
    #   requests
cloud-sql-python-connector[pg8000]==1.2.1
    # via mpcq (setup.py)
cryptography==39.0.2
    # via cloud-sql-python-connector
distlib==0.3.6
    # via virtualenv
exceptiongroup==1.1.1
    # via pytest
filelock==3.10.0
    # via virtualenv
frozenlist==1.3.3
    # via
    #   aiohttp
    #   aiosignal
google-api-core[grpc]==2.11.0
    # via google-cloud-secret-manager
google-auth==2.16.2
    # via
    #   cloud-sql-python-connector
    #   google-api-core
google-cloud-secret-manager==2.16.0
    # via mpcq (setup.py)
googleapis-common-protos[grpc]==1.59.0
    # via
    #   google-api-core
    #   grpc-google-iam-v1
    #   grpcio-status
grpc-google-iam-v1==0.12.6
    # via google-cloud-secret-manager
grpcio==1.51.3
    # via
    #   google-api-core
    #   googleapis-common-protos
    #   grpc-google-iam-v1
    #   grpcio-status
grpcio-status==1.51.3
    # via google-api-core
identify==2.5.21
    # via pre-commit
idna==3.4
    # via
    #   requests
    #   yarl
iniconfig==2.0.0
    # via pytest
multidict==6.0.4
    # via
    #   aiohttp
    #   yarl
nodeenv==1.7.0
    # via pre-commit
numpy==1.24.2
    # via
    #   astropy
    #   pandas
    #   pyerfa
packaging==23.0
    # via
    #   astropy
    #   pytest
pandas==1.5.3
    # via mpcq (setup.py)
pg8000==1.29.4
    # via
    #   cloud-sql-python-connector
    #   mpcq (setup.py)
platformdirs==3.1.1
    # via virtualenv
pluggy==1.0.0
    # via pytest
pre-commit==3.2.0
    # via mpcq (setup.py)
proto-plus==1.22.2
    # via google-cloud-secret-manager
protobuf==4.22.1
    # via
    #   google-api-core
    #   google-cloud-secret-manager
    #   googleapis-common-protos
    #   grpc-google-iam-v1
    #   grpcio-status
    #   proto-plus
pyasn1==0.4.8
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.2.8
    # via google-auth
pycparser==2.21
    # via cffi
pyerfa==2.0.0.2
    # via astropy
pytest==7.2.2
    # via mpcq (setup.py)
python-dateutil==2.8.2
    # via
    #   pandas
    #   pg8000
pytz==2022.7.1
    # via pandas
pyyaml==6.0
    # via
    #   astropy
    #   pre-commit
requests==2.28.2
    # via
    #   cloud-sql-python-connector
    #   google-api-core
rsa==4.9
    # via google-auth
scramp==1.4.4
    # via pg8000
six==1.16.0
    # via
    #   google-auth
    #   python-dateutil
sqlalchemy==1.4.47
    # via mpcq (setup.py)
tomli==2.0.1
    # via pytest
urllib3==1.26.15
    # via requests
virtualenv==20.21.0
    # via pre-commit
yarl==1.8.2
    # via aiohttp

# The following packages are considered to be unsafe in a requirements file:
# setuptools
